project_legacy_backend/requirements.txt
django-allauth
djangorestframework
django-cors-headers
requests
PyJWT
cryptography